# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reactions:
    get:
      operationId: listReactions
      tags: [reactions]
      summary: List all reactions
      responses:
        "200":
          description: List of reactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reaction"
    post:
      operationId: createReaction
      tags: [reactions]
      summary: Create a new reaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReactionBody"
      responses:
        "201":
          description: Created reaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
  /reactions/{id}:
    get:
      operationId: getReaction
      tags: [reactions]
      summary: Get a reaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateReaction
      tags: [reactions]
      summary: Update a reaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReactionBody"
      responses:
        "200":
          description: Updated reaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteReaction
      tags: [reactions]
      summary: Delete a reaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reactions/{id}/candidates:
    get:
      operationId: listCandidatesForReaction
      tags: [candidates]
      summary: List candidates for a reaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sortBy
          in: query
          required: false
          description: Field to sort candidates by
          schema:
            type: string
            enum:
              - rank
              - predictedActivity
              - predictedSelectivity
              - predictedStability
              - confidenceScore
        - name: sortDir
          in: query
          required: false
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
        - name: source
          in: query
          required: false
          description: Filter by candidate source
          schema:
            type: string
            enum: [generated, literature]
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
  /reactions/{id}/generate-candidates:
    post:
      operationId: generateCandidates
      tags: [candidates]
      summary: AI-generate novel candidate designs for a reaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCandidatesBody"
      responses:
        "200":
          description: Generated candidates saved and returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
  /reactions/{id}/search-candidates:
    post:
      operationId: searchCandidates
      tags: [candidates]
      summary: Search PubChem and ChEMBL for known literature candidates and persist them
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Persisted literature candidates with cheminformatics metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchCandidatesResponse"
        "404":
          description: Reaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Reaction:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        equation:
          type: string
        targetProduct:
          type: string
        conditions:
          type: string
        description:
          type: string
        domain:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - equation
        - targetProduct
        - conditions
        - description
        - domain
        - createdAt
        - updatedAt
    CreateReactionBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        equation:
          type: string
        targetProduct:
          type: string
        conditions:
          type: string
        description:
          type: string
        domain:
          type: string
      required:
        - name
        - type
        - equation
        - targetProduct
        - conditions
        - description
        - domain
    UpdateReactionBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        equation:
          type: string
        targetProduct:
          type: string
        conditions:
          type: string
        description:
          type: string
        domain:
          type: string
    Candidate:
      type: object
      properties:
        id:
          type: integer
        reactionId:
          type: integer
        name:
          type: string
        formula:
          type: string
        source:
          type: string
        sourceDb:
          type: ["string", "null"]
        candidateType:
          type: ["string", "null"]
        routeType:
          type: ["string", "null"]
        predictedActivity:
          type: number
        predictedSelectivity:
          type: number
        predictedStability:
          type: number
        confidenceScore:
          type: number
        feedstockFitScore:
          type: ["number", "null"]
        costScore:
          type: ["number", "null"]
        sustainabilityScore:
          type: ["number", "null"]
        scalabilityScore:
          type: ["number", "null"]
        uncertaintyScore:
          type: ["number", "null"]
        mechanismText:
          type: string
        structureData:
          type: string
        evidenceText:
          type: ["string", "null"]
        energyProfileData:
          type: ["string", "null"]
        pathwayData:
          type: ["string", "null"]
        rank:
          type: ["integer", "null"]
        pubchemCid:
          type: ["integer", "null"]
        chemblId:
          type: ["string", "null"]
        molecularWeight:
          type: ["number", "null"]
        logP:
          type: ["number", "null"]
        tpsa:
          type: ["number", "null"]
        canonicalSmiles:
          type: ["string", "null"]
        iupacName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - reactionId
        - name
        - formula
        - source
        - predictedActivity
        - predictedSelectivity
        - predictedStability
        - confidenceScore
        - mechanismText
        - structureData
        - createdAt
        - updatedAt
    SearchCandidatesResponse:
      type: object
      properties:
        query:
          type: string
        sourcesQueried:
          type: array
          items:
            type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
      required:
        - query
        - sourcesQueried
        - candidates
...
    GenerateCandidatesBody:
      type: object
      properties:
        count:
          type: integer
      required:
        - count
...